+++ clover/chrony.conf
@@ -1,0 +1,15 @@
+#Serve time based on the local system clock
+#tells Chrony to act as an NTP server
+local stratum 8
+
+#Allow clients on your local subnet to query the server
+allow 192.168.0.0/24
+
+# Optional: Bind Chrony to the specific Ethernet interface
+# This prevents it from trying to serve time over Wi-Fi or other adapters.
+bindaddress 192.168.0.150
+
+# 5. Standard system files
+driftfile /var/lib/chrony/chrony.drift
+makestep 1 3
+rtcsync
