+++ clover/prj.conf
@@ -8,13 +8,13 @@ CONFIG_NEWLIB_LIBC=y
 CONFIG_CBPRINTF_LIBC_SUBSTS=y
 
 # Set the following if you're connected directly via ethernet.
-CONFIG_NET_CONFIG_MY_IPV4_ADDR="169.254.99.99"
-CONFIG_NET_CONFIG_MY_IPV4_NETMASK="255.255.0.0"
+#CONFIG_NET_CONFIG_MY_IPV4_ADDR="169.254.99.99"
+#CONFIG_NET_CONFIG_MY_IPV4_NETMASK="255.255.0.0"
 
 # Set the following if you're connected via the DAQ box router.
-# CONFIG_NET_CONFIG_MY_IPV4_ADDR="192.168.0.150"
-# CONFIG_NET_CONFIG_MY_IPV4_NETMASK="255.255.255.0"
-# CONFIG_NET_CONFIG_MY_IPV4_GW="192.168.0.1"
+CONFIG_NET_CONFIG_MY_IPV4_ADDR="192.168.0.150"
+CONFIG_NET_CONFIG_MY_IPV4_NETMASK="255.255.255.0"
+CONFIG_NET_CONFIG_MY_IPV4_GW="192.168.0.1"
 
 # Add nanopb, used to serialize and deserialize requests and responses for the command server.
 CONFIG_NANOPB=y
@@ -47,7 +47,7 @@ CONFIG_NET_MAX_CONTEXTS=64
 CONFIG_NET_CONNECTION_MANAGER=y
 CONFIG_NET_TCP_TIME_WAIT_DELAY=0
 CONFIG_POSIX_NETWORKING=y  # Links POSIX networking functions like inet_addr
-
+CONFIG_SNTP=y
 # Enable ADC and logging
 CONFIG_ADC=y
 CONFIG_ADC_LOG_LEVEL_WRN=y
@@ -61,9 +61,3 @@ CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=4096
 # CONFIG_THREAD_ANALYZER_AUTO=y
 # CONFIG_THREAD_ANALYZER_AUTO_INTERVAL=5
 # CONFIG_THREAD_NAME=y
-
-#NTP Servers
-server 0.us.pool.ntp.org
-server 1.us.pool.ntp.org
-server 2.us.pool.ntp.org
-server 3.us.pool.ntp.org
